requirements.txt
google-adk
python-dotenv
litellm
pydantic
finnhub-python